Test that empty ELF sections have the correct file offset
